# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Expense Tracker API
  version: 1.0.0
  description: Local REST API for the Personal Expense Tracker
  license:
    name: MIT
servers:
- url: http://localhost:3000
  description: Local backend
components:
  schemas:
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum:
          - expense
          - income
        date:
          type: string
          format: date
        amountCents:
          type: integer
        categoryId:
          type: integer
          nullable: true
        incomeSourceId:
          type: integer
          nullable: true
        paymentMethodId:
          type: integer
        note:
          type: string
          maxLength: 500
          nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        attachmentId:
          type: string
          nullable: true
        recurringRuleId:
          type: integer
          nullable: true
        occurrenceDate:
          type: string
          format: date
          nullable: true
      required:
      - id
      - type
      - date
      - amountCents
      - paymentMethodId
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
          nullable: true
        sortOrder:
          type: integer
        archived:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - name
      - sortOrder
      - archived
    TopExpenseItem:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        amountCents:
          type: integer
        categoryId:
          type: integer
        note:
          type: string
          nullable: true
      required:
      - id
      - date
      - amountCents
      - categoryId
...
    RecurringRule:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum:
          - expense
          - income
        amountCents:
          type: integer
        categoryId:
          type: integer
          nullable: true
        incomeSourceId:
          type: integer
          nullable: true
        paymentMethodId:
          type: integer
        frequency:
          type: string
          enum:
          - weekly
          - monthly
          - yearly
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - type
      - amountCents
      - paymentMethodId
      - frequency
      - startDate
      - active
...
paths:
...
  /api/dashboard/years:
    get:
      summary: Available Years
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  years:
                    type: array
                    items:
                      type: object
                      properties:
                        year:
                          type: integer
                        hasData:
                          type: boolean
                        totals:
                          type: object
                          properties:
                            incomeCents:
                              type: integer
                            expenseCents:
                              type: integer
                            netCents:
                              type: integer
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: getDashboardYears
  /api/dashboard/annual:
    get:
      summary: Annual dashboard
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  totals:
                    type: object
                    properties:
                      incomeCents:
                        type: integer
                      expenseCents:
                        type: integer
                      netCents:
                        type: integer
                  monthly:
                    type: array
                    items:
                      type: object
                      properties:
                        month:
                          type: integer
                        incomeCents:
                          type: integer
                        expenseCents:
                          type: integer
                        netCents:
                          type: integer
                  breakdowns:
                    type: object
                    properties:
                      byCategory:
                        type: array
                        items:
                          type: object
                      byIncomeSource:
                        type: array
                        items:
                          type: object
                      byPaymentMethod:
                        type: array
                        items:
                          type: object
                      byTag:
                        type: array
                        items:
                          type: object
                  topExpenses:
                    type: array
                    items:
                      $ref: '#/components/schemas/TopExpenseItem'
                  isPartialPeriod:
                    type: boolean
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: getDashboardAnnual
  /api/dashboard/annual/compare:
    get:
      summary: Year comparison
      parameters:
      - name: yearA
        in: query
        required: true
        schema:
          type: integer
      - name: yearB
        in: query
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  yearA:
                    type: object
                  yearB:
                    type: object
                  delta:
                    type: object
                    properties:
                      incomePct:
                        type: number
                        nullable: true
                      expensePct:
                        type: number
                        nullable: true
                      netPct:
                        type: number
                        nullable: true
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: getDashboardAnnualCompare
  /api/dashboard/monthly:
    get:
      summary: Monthly dashboard
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
      - name: month
        in: query
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  totals:
                    type: object
                  breakdowns:
                    type: object
                  topExpenses:
                    type: array
                    items:
                      $ref: '#/components/schemas/TopExpenseItem'
                  isPartialPeriod:
                    type: boolean
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: getDashboardMonthly
...
  /api/import/csv/preview:
    post:
      summary: Upload + preview
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileToken:
                    type: string
                  headers:
                    type: array
                    items:
                      type: string
                  sampleRows:
                    type: array
                    items:
                      type: array
                      items:
                        type: string
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: postImportCsvPreview
  /api/import/csv/commit:
    post:
      summary: Start commit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fileToken:
                  type: string
                mapping:
                  type: object
                  properties:
                    date:
                      type: string
                    note:
                      type: string
                    amount:
                      type: string
                    type:
                      type: string
                    signedAmount:
                      type: string
                defaultCategoryId:
                  type: integer
                defaultIncomeSourceId:
                  type: integer
                defaultPaymentMethodId:
                  type: integer
              required:
              - fileToken
              - mapping
              - defaultPaymentMethodId
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
        '400':
          description: Bad Request
        '410':
          description: Gone
        '428':
          description: Precondition Required
      operationId: postImportCsvCommit
  /api/import/csv/commit/{jobId}:
    parameters:
    - name: jobId
      in: path
      required: true
      schema:
        type: string
    get:
      summary: Poll commit progress
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum:
                    - processing
                    - done
                  rowsProcessed:
                    type: integer
                  totalRows:
                    type: integer
                  imported:
                    type: integer
                  skipped:
                    type: array
                    items:
                      type: object
                      properties:
                        row:
                          type: integer
                        reason:
                          type: string
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: getImportCsvCommitJobid
...
  /api/recurring-rules:
    get:
      summary: List recurring rules
      parameters:
      - name: includeStopped
        in: query
        schema:
          type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringRule'
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: getRecurring-rules
    post:
      summary: Create recurring rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum:
                  - expense
                  - income
                amountCents:
                  type: integer
                categoryId:
                  type: integer
                incomeSourceId:
                  type: integer
                paymentMethodId:
                  type: integer
                frequency:
                  type: string
                  enum:
                  - weekly
                  - monthly
                  - yearly
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
              required:
              - type
              - amountCents
              - paymentMethodId
              - frequency
              - startDate
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringRule'
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: postRecurring-rules
  /api/recurring-rules/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    put:
      summary: Update recurring rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum:
                  - expense
                  - income
                amountCents:
                  type: integer
                categoryId:
                  type: integer
                incomeSourceId:
                  type: integer
                paymentMethodId:
                  type: integer
                frequency:
                  type: string
                  enum:
                  - weekly
                  - monthly
                  - yearly
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringRule'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: putRecurring-rulesId
  /api/recurring-rules/{id}/stop:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    post:
      summary: Stop recurring rule
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringRule'
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: postRecurring-rulesIdStop
  /api/recurring-rules/{id}/resume:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    post:
      summary: Resume recurring rule
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringRule'
        '404':
          description: Not Found
        '428':
          description: Precondition Required
      operationId: postRecurring-rulesIdResume
  /api/recurring-rules/generate:
    post:
      summary: On-demand generation check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad Request
        '428':
          description: Precondition Required
      operationId: postRecurring-rulesGenerate
...
security: []